+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 * Fixed full screen safe area issue in cupertino controls.
 * Fixed subtitles duplication after changing data source.
 * [BREAKING_CHANGE] Changed min. Flutter version to 2.2.3.
+* Added url parameter for changedResolution event.
 
 ## 0.0.76
 * Fixed iOS build issue.